+++ L1Trigger/L1TMuon/interface/deprecate/DTBunchCrossingCleaner.h
@@ -10,7 +10,7 @@
 // Author: L. Gray (FNAL)
 //
 #include <vector>
-#include "L1Trigger/L1TMuon/interface/deprecate/MuonTriggerPrimitiveFwd.h"
+#include "L1Trigger/L1TMuon/interface/MuonTriggerPrimitiveFwd.h"
 #include "FWCore/Utilities/interface/InputTag.h"
 
 namespace edm{
