+++ L1Trigger/L1TMuonBarrel/src/Twinmux_v1/DTBunchCrossingCleaner.h
@@ -10,7 +10,7 @@
 // Author: L. Gray (FNAL)
 //
 #include <vector>
-#include "L1Trigger/L1TMuon/interface/deprecate/MuonTriggerPrimitive.h"
+#include "L1Trigger/L1TMuon/interface/MuonTriggerPrimitive.h"
 #include "FWCore/Utilities/interface/InputTag.h"
 
 namespace edm{
