+++ L1Trigger/L1TMuonBarrel/src/Twinmux_v1/MBLTProducer.cc
@@ -9,7 +9,7 @@
 #include "FWCore/Framework/interface/MakerMacros.h"
 #include <FWCore/Framework/interface/ConsumesCollector.h>
 
-#include "L1Trigger/L1TMuon/interface/deprecate/MuonTriggerPrimitive.h"
+#include "L1Trigger/L1TMuon/interface/MuonTriggerPrimitive.h"
 #include "L1Trigger/L1TMuonBarrel/src/Twinmux_v1/MBLTCollection.h"
 #include "L1Trigger/L1TMuonBarrel/src/Twinmux_v1/MBLTCollectionFwd.h"
