+++ L1Trigger/L1TMuonBarrel/src/Twinmux_v1/PrimitiveCombiner.h
@@ -10,7 +10,7 @@
 //
 
 #include "FWCore/Framework/interface/ESHandle.h"
-#include "L1Trigger/L1TMuon/interface/deprecate/MuonTriggerPrimitive.h"
+#include "L1Trigger/L1TMuon/interface/MuonTriggerPrimitive.h"
 
 #include <cmath>
