+++ L1Trigger/L1TMuonBarrel/src/Twinmux_v1/RPCCollector.h
@@ -7,7 +7,7 @@
 //       Positional information is not assigned here.
 //
 #include <vector>
-#include "L1Trigger/L1TMuon/interface/deprecate/MuonTriggerPrimitive.h"
+#include "L1Trigger/L1TMuon/interface/MuonTriggerPrimitive.h"
 #include "FWCore/Utilities/interface/InputTag.h"
 #include "DataFormats/RPCDigi/interface/RPCDigi.h"
 #include "DataFormats/RPCDigi/interface/RPCDigiCollection.h"
