+++ L1Trigger/L1TMuonEndCap/interface/EmulatorClasses.h
@@ -2,7 +2,7 @@
 #define __EmulatorClasses_
 
 #include "L1Trigger/L1TMuonEndCap/interface/PhiMemoryImage.h"
-#include "L1Trigger/L1TMuon/interface/deprecate/MuonTriggerPrimitive.h"
+#include "L1Trigger/L1TMuon/interface/MuonTriggerPrimitive.h"
 
 typedef std::vector<std::vector<PhiMemoryImage>> ImageCollector;
 typedef std::vector<std::vector<int>> Code;
